Bug 25954: (follow-up) Correct label "for" attribute
This patch updates a couple of search include forms to change the "for"
attribute of the label. Although "idx" is the first field in this
situation, the label's text applies to the keyword field so it should
give focus to the keyword field.

Signed-off-by: Josef Moravec <josef.moravec@gmail.com>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
(cherry picked from commit 1920c92d811dc5477a9cb1064f6f7e2dd473a380)
Signed-off-by: Fridolin Somers <fridolin.somers@biblibre.com>