Bug 16357: (QA follow-up) Add log4perl configs during package upgrade
If plack.psgi is updated to the newer version and the log4perl.conf file
is not then the warnings will not be logged anywhere. This adds the
log4perl configurations that are needed for logging for pre-existing Koha
installation which are upgraded.

Signed-off-by: Arthur Suzuki <arthur.suzuki@biblibre.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
(cherry picked from commit 5651facf0574f028250aeb8f30d6d8e2a5ebaabe)

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>